requirements.txt
gradio==3.27.0
matplotlib==3.7.1
networkx==2.8.4
numpy==1.24.3
pandas==1.5.3
Pillow==9.3.0
pyro_ppl==1.8.4
scikit_image==0.20.0
tqdm==4.65.0
torchvision
torchxrayvision
scikit-learn
python-ternary